# docs/swagger.json
{
    "components": {
        "schemas": {
            "api.ErrorResponse": {
                "properties": {
                    "error": {
                        "example": "validation failed",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "model.SDKInstance": {
                "properties": {
                    "address": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "last_poll_at": {
                        "type": "string"
                    },
                    "registered_at": {
                        "type": "string"
                    },
                    "service": {
                        "type": "string"
                    },
                    "status": {
                        "description": "\"alive\",\"stale\",\"dead\" \u2014 computed, not persisted",
                        "type": "string"
                    },
                    "version": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "info": {
        "description": "Central coordination controller for the atropos ecosystem.\nManages rules, faults, experiments, workflows, and SDK lifecycle.",
        "title": "Manteion Control-Plane API",
        "version": "1.0"
    },
    "paths": {
...
        "/sdk/instances": {
            "get": {
                "description": "Returns all currently registered SDK instances. Status is computed\n(alive/stale/dead) from last-poll-at.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/model.SDKInstance"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "List SDK instances",
                "tags": [
                    "sdk"
                ]
            }
        },
        "/sdk/register": {
            "post": {
                "description": "Registers an atropos-go SDK process. The 201 response carries\nstatus=\"registered\" and may include initial intent payload\n(rules, active_fault, freeze_cfg) when an IntentReader is\nconfigured for the service.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/model.SDKInstance",
                                        "summary": "instance",
                                        "description": "SDK instance metadata"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "SDK instance metadata",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {},
                                    "type": "object"
                                }
                            }
                        },
                        "description": "registration ack with optional intent fields"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "invalid JSON or registration error"
                    }
                },
                "summary": "Register SDK instance",
                "tags": [
                    "sdk"
                ]
            }
        },
        "/sdk/register/{id}": {
            "delete": {
                "parameters": [
                    {
                        "description": "Instance ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "instance deregistered"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "instance not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Deregister SDK instance",
                "tags": [
                    "sdk"
                ]
            }
        },
        "/sdk/rules": {
            "get": {
                "description": "SDKs send their last-known rule version via the version query param.\nReturns 304 if unchanged, 200 with the current rule set otherwise.\nThe 200 body wraps the version and the compiled rule list:\n{\"version\": uint64, \"rules\": []ruleconv.CompiledRule}.\nOptional instance_id query param triggers a best-effort poll-timestamp touch.",
                "parameters": [
                    {
                        "description": "service name",
                        "in": "query",
                        "name": "service",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "last known rule version (uint64); omit on first poll",
                        "in": "query",
                        "name": "version",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "SDK instance ID for poll-timestamp tracking",
                        "in": "query",
                        "name": "instance_id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {},
                                    "type": "object"
                                }
                            }
                        },
                        "description": "{version, rules: []ruleconv.CompiledRule}"
                    },
                    "304": {
                        "description": "no rule changes since requested version"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "service query parameter required"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Poll for rule updates",
                "tags": [
                    "sdk"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "description": "Local dev (HTTP)",
            "url": "http://localhost:8080/api/v1"
        },
        {
            "description": "Local dev (HTTPS)",
            "url": "https://localhost:8080/api/v1"
        }
    ]
}